+++ griptape/artifacts/__init__.py
@@ -1,30 +1,28 @@
 from .base_artifact import BaseArtifact
-from .error_artifact import ErrorArtifact
-from .info_artifact import InfoArtifact
+from .base_system_artifact import BaseSystemArtifact
+
 from .text_artifact import TextArtifact
-from .json_artifact import JsonArtifact
 from .blob_artifact import BlobArtifact
-from .boolean_artifact import BooleanArtifact
-from .csv_row_artifact import CsvRowArtifact
-from .list_artifact import ListArtifact
-from .media_artifact import MediaArtifact
 from .image_artifact import ImageArtifact
 from .audio_artifact import AudioArtifact
+from .json_artifact import JsonArtifact
 from .action_artifact import ActionArtifact
 from .generic_artifact import GenericArtifact
+
+from .error_artifact import ErrorArtifact
+from .info_artifact import InfoArtifact
+from .list_artifact import ListArtifact
 
 
 __all__ = [
     "BaseArtifact",
+    "BaseSystemArtifact",
     "ErrorArtifact",
     "InfoArtifact",
     "TextArtifact",
     "JsonArtifact",
     "BlobArtifact",
-    "BooleanArtifact",
-    "CsvRowArtifact",
     "ListArtifact",
-    "MediaArtifact",
     "ImageArtifact",
     "AudioArtifact",
     "ActionArtifact",