+++ griptape/artifacts/base_artifact.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 import uuid
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, Optional
@@ -15,6 +14,18 @@ if TYPE_CHECKING:
 
 @define
 class BaseArtifact(SerializableMixin, ABC):
+    """Serves as the base class for all Artifacts.
+
+    Artifacts are used to store data that can be provided as input to, or received as output from, a language model (LLM).
+
+    Attributes:
+        id: The unique identifier of the Artifact. Defaults to a random UUID.
+        reference: The optional Reference to the Artifact.
+        meta: The metadata associated with the Artifact. Defaults to an empty dictionary.
+        name: The name of the Artifact. Defaults to the id.
+        value: The value of the Artifact.
+    """
+
     id: str = field(default=Factory(lambda: uuid.uuid4().hex), kw_only=True, metadata={"serializable": True})
     reference: Optional[Reference] = field(default=None, kw_only=True, metadata={"serializable": True})
     meta: dict[str, Any] = field(factory=dict, kw_only=True, metadata={"serializable": True})
@@ -24,22 +35,7 @@ class BaseArtifact(SerializableMixin, ABC):
         metadata={"serializable": True},
     )
     value: Any = field()
-
-    @classmethod
-    def value_to_bytes(cls, value: Any) -> bytes:
-        if isinstance(value, bytes):
-            return value
-        else:
-            return str(value).encode()
-
-    @classmethod
-    def value_to_dict(cls, value: Any) -> dict:
-        dict_value = value if isinstance(value, dict) else json.loads(value)
-
-        return dict(dict_value.items())
-
-    def to_text(self) -> str:
-        return str(self.value)
+    encoding: str = field(default="utf-8", kw_only=True)
 
     def __str__(self) -> str:
         return self.to_text()
@@ -50,5 +46,8 @@ class BaseArtifact(SerializableMixin, ABC):
     def __len__(self) -> int:
         return len(self.value)
 
+    def to_bytes(self) -> bytes:
+        return self.to_text().encode()
+
     @abstractmethod
-    def __add__(self, other: BaseArtifact) -> BaseArtifact: ...
+    def to_text(self) -> str: ...