+++ griptape/artifacts/json_artifact.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import json
-from typing import Union
+from typing import Any, Union
 
 from attrs import define, field
 
@@ -10,12 +10,22 @@ from griptape.artifacts import BaseArtifact
 Json = Union[dict[str, "Json"], list["Json"], str, int, float, bool, None]
 
 
+def value_to_json(value: Any) -> Json:
+    if isinstance(value, str):
+        return json.loads(value)
+    else:
+        return json.loads(json.dumps(value))
+
+
 @define
 class JsonArtifact(BaseArtifact):
-    value: Json = field(converter=lambda v: json.loads(json.dumps(v)), metadata={"serializable": True})
+    """Stores JSON data.
+
+    Attributes:
+        value: The JSON data. Values will automatically be converted to a JSON-compatible format.
+    """
+
+    value: Json = field(converter=value_to_json, metadata={"serializable": True})
 
     def to_text(self) -> str:
         return json.dumps(self.value)
-
-    def __add__(self, other: BaseArtifact) -> JsonArtifact:
-        raise NotImplementedError