+++ griptape/artifacts/text_artifact.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Any, Optional
 
 from attrs import define, field
 
@@ -11,22 +11,35 @@ if TYPE_CHECKING:
     from griptape.tokenizers import BaseTokenizer
 
 
+def value_to_str(value: Any) -> str:
+    if isinstance(value, dict):
+        return "\n".join(f"{key}: {val}" for key, val in value.items())
+    else:
+        return str(value)
+
+
 @define
 class TextArtifact(BaseArtifact):
-    value: str = field(converter=str, metadata={"serializable": True})
+    value: str = field(converter=value_to_str, metadata={"serializable": True})
     encoding: str = field(default="utf-8", kw_only=True)
     encoding_error_handler: str = field(default="strict", kw_only=True)
     _embedding: list[float] = field(factory=list, kw_only=True)
-
-    @property
-    def embedding(self) -> Optional[list[float]]:
-        return None if len(self._embedding) == 0 else self._embedding
 
     def __add__(self, other: BaseArtifact) -> TextArtifact:
         return TextArtifact(self.value + other.value)
 
     def __bool__(self) -> bool:
         return bool(self.value.strip())
+
+    @property
+    def embedding(self) -> Optional[list[float]]:
+        return None if len(self._embedding) == 0 else self._embedding
+
+    def to_text(self) -> str:
+        return self.value
+
+    def to_bytes(self) -> bytes:
+        return str(self.value).encode(encoding=self.encoding, errors=self.encoding_error_handler)
 
     def generate_embedding(self, driver: BaseEmbeddingDriver) -> Optional[list[float]]:
         self._embedding.clear()
@@ -36,6 +49,3 @@ class TextArtifact(BaseArtifact):
 
     def token_count(self, tokenizer: BaseTokenizer) -> int:
         return tokenizer.count_tokens(str(self.value))
-
-    def to_bytes(self) -> bytes:
-        return str(self.value).encode(encoding=self.encoding, errors=self.encoding_error_handler)