+++ griptape/engines/extraction/csv_extraction_engine.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Optional, cast
 
 from attrs import Factory, define, field
 
-from griptape.artifacts import CsvRowArtifact, ListArtifact, TextArtifact
+from griptape.artifacts import ListArtifact, TextArtifact
 from griptape.common import Message, PromptStack
 from griptape.engines import BaseExtractionEngine
 from griptape.utils import J2
@@ -36,22 +36,22 @@ class CsvExtractionEngine(BaseExtractionEngine):
             item_separator="\n",
         )
 
-    def text_to_csv_rows(self, text: str, column_names: list[str]) -> list[CsvRowArtifact]:
+    def text_to_csv_rows(self, text: str, column_names: list[str]) -> list[TextArtifact]:
         rows = []
 
         with io.StringIO(text) as f:
             for row in csv.reader(f):
-                rows.append(CsvRowArtifact(dict(zip(column_names, [x.strip() for x in row]))))
+                rows.append(TextArtifact(dict(zip(column_names, [x.strip() for x in row]))))
 
         return rows
 
     def _extract_rec(
         self,
         artifacts: list[TextArtifact],
-        rows: list[CsvRowArtifact],
+        rows: list[TextArtifact],
         *,
         rulesets: Optional[list[Ruleset]] = None,
-    ) -> list[CsvRowArtifact]:
+    ) -> list[TextArtifact]:
         artifacts_text = self.chunk_joiner.join([a.value for a in artifacts])
         system_prompt = self.system_template_generator.render(
             column_names=self.column_names,