+++ tests/unit/artifacts/test_audio_artifact.py
@@ -6,20 +6,22 @@ from griptape.artifacts import AudioArtifact, BaseArtifact
 class TestAudioArtifact:
     @pytest.fixture()
     def audio_artifact(self):
-        return AudioArtifact(value=b"some binary audio data", format="pcm", model="provider/model", prompt="two words")
+        return AudioArtifact(
+            value=b"some binary audio data", format="pcm", meta={"model": "provider/model", "prompt": "two words"}
+        )
 
     def test_mime_type(self, audio_artifact: AudioArtifact):
         assert audio_artifact.mime_type == "audio/pcm"
 
     def test_to_text(self, audio_artifact: AudioArtifact):
-        assert audio_artifact.to_text() == "Media, type: audio/pcm, size: 22 bytes"
+        assert audio_artifact.to_text() == "Audio, format: pcm, size: 22 bytes"
 
     def test_to_dict(self, audio_artifact: AudioArtifact):
         audio_dict = audio_artifact.to_dict()
 
         assert audio_dict["format"] == "pcm"
-        assert audio_dict["model"] == "provider/model"
-        assert audio_dict["prompt"] == "two words"
+        assert audio_dict["meta"]["model"] == "provider/model"
+        assert audio_dict["meta"]["prompt"] == "two words"
         assert audio_dict["value"] == "c29tZSBiaW5hcnkgYXVkaW8gZGF0YQ=="
 
     def test_deserialization(self, audio_artifact):
@@ -31,5 +33,5 @@ class TestAudioArtifact:
         assert deserialized_artifact.value == b"some binary audio data"
         assert deserialized_artifact.mime_type == "audio/pcm"
         assert deserialized_artifact.format == "pcm"
-        assert deserialized_artifact.model == "provider/model"
-        assert deserialized_artifact.prompt == "two words"
+        assert deserialized_artifact.meta["model"] == "provider/model"
+        assert deserialized_artifact.meta["prompt"] == "two words"